Move add_overviews() to trollflow_sat.utils
